ZOOKEEPER-4970: Deprecate ZKConfig methods which throw QuorumPeerConfig.ConfigException
ZOOKEEPER-4970: Deprecate ZKConfig methods which throw QuorumPeerConfig.ConfigException
`QuorumPeerConfig` belongs to and ties with other server code. To
separate a slimmer `zookeeper-client` jar, we have to break the
dependency from `ZKConfig` to `QuorumPeerConfig`. But this is a breaking
change, it would be good to deprecate these methods first to ease future
migration.
Refs: ZOOKEEPER-4966, ZOOKEEPER-4970, ZOOKEEPER-233.
Reviewers: tisonkun
Author: kezhuw
Closes #2309 from kezhuw/ZOOKEEPER-4970-deprecate-ConfigException-from-ZKConfig